requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
keras-image-helper==0.0.1
numpy==1.24.1
Pillow==9.4.0
requests==2.28.2
tflite-runtime==2.11.0
urllib3==1.26.14
